+++ EssentialFeed/EssentialFeed/Feed API/RemoteFeedLoader.swift
@@ -7,15 +7,6 @@
 
 import Foundation
 
-public enum HTTPClientResult {
-    case success(Data, HTTPURLResponse)
-    case failure(Error)
-}
-
-public protocol HTTPClient {
-    func get(from url: URL, completion: @escaping (HTTPClientResult) -> Void)
-}
-
 public final class RemoteFeedLoader {
     private let url: URL
     private let client: HTTPClient
@@ -39,47 +30,14 @@ public final class RemoteFeedLoader {
         client.get(from: url) { result in
             switch result {
             case .success(let data, let response):
-                do {
-                    let feedItems = try FeedItemsMapper.map(data, response)
-                    completion(.success(feedItems))
-                } catch {
-                    completion(.failure(.invalidData))
-                }
+                completion(self.map(data: data, response: response))
             case .failure:
                 completion(.failure(.connectivity))
             }
         }
     }
-}
 
-final private class FeedItemsMapper {
-    private static var STATUS_OK = 200
-
-    static func map(_ data: Data, _ response: HTTPURLResponse) throws -> [FeedItem] {
-        guard response.statusCode == STATUS_OK else {
-            throw RemoteFeedLoader.Error.invalidData
-        }
-        let root = try JSONDecoder().decode(Root.self, from: data)
-        return root.items.map { $0.feedItem }
-    }
-}
-
-private struct Root: Decodable {
-    let items: [Item]
-}
-
-private struct Item: Decodable {
-    let id: UUID
-    let description: String?
-    let location: String?
-    let image: URL
-
-    var feedItem: FeedItem {
-        FeedItem(
-            id: id,
-            description: description,
-            location: location,
-            imageURL: image
-        )
+    func map(data: Data, response: HTTPURLResponse) -> Result {
+        FeedItemsMapper.map(data, response)
     }
 }
